backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
dnspython
redis
python-dotenv
aiofiles
loguru
tenacity
dnspython
